move formatComment to separate file
